Update API URL for SmokeTest
Actually not sure if this will resolve properly from within the build project
but it's a quick change so worth trying. There are going to be other changes
required on the data side because we've switched to a completely new API
instance but this could be all we need for code changes.