DEV-2296 Fix intermittent failure in batch
The symptoms of the problem were that the "finished" counter in the batches
was incrementing rather quickly but this did not appear to be tied to actual VM
lifecycles.

Turns out this was due to failing to consider the API in the creation of the
VM configuration. The inputs were being resolved against the dataset but the
traffic on the API was too much and it was falling over.

I've introduced retry mechanisms on the API communication and also am now
properly handling failures such that exceptions outside
of the ComputeEngine are now rippled back up to the status reporter as
pipeline failures.

Also correct typo I happened across and commit the tunings I did to the
TeloBatch.