DEV-2303 Make batch defaults safer
The batch framework currently makes it a bit too easy to create expensive
runs. This change removes the ability to use non-pre-emptible VMs, as the cost
difference is a factor of 5x.

Also change the default for VM definitions to a custom machine with 8 cores
and 8GB RAM. The references in the rest of the pipeline code have been updated
to use custom machines reflecting the original default.