DEV-2203 Add optionals to orange inputs to support shallow
Linx was the root of the problem, but also the cuppa, peach and virus were inconsistent with the
other stages handling of skipped output.

That said, setting the output to empty in the skipped output method is probably a better solution,
and we should come back and change all staged to behave that way.