DEV-2268 Override resources per-directory
This changes the behaviour of the private resources override such that
rather than simply ovewriting files from the public repository with
those from the private that have the same path, the parent directories
will be cleared out too.

This means that rather than relying on filenames to be identical to
avoid having public resources bleed through to the private image, we
will now avoid that problem altogether. However the private resources
need to always have complete self-contained resource sets.

Note that this implementation assumes there will not be complex
multi-level overrides: all resources should be in a "leaf" directory or
imaging is likely to fail.