DEV-2188 Tag public git repo on image completion
These commands are placed at the very end and everything after them
should succeed. The tag will match the image name, which consists of the
pipeline version and a timestamp.

The checkout is gelded after the tagging is pushed so as to avoid the
possibility of anything behing pushed accidentally from a VM created
from the image.

Also place what appears to be a redundant `apt update` before the
pulling down of the packages needed for adding the CRAN repository. This
was required as one of the dependencies of the `ca-certificates` package
began to fail to install today as it had been updated upstream.