DEV-2188 Remove private imager external IP
There were some problems this morning with the clone of the private
resources repository. I don't know whether this may have been rectified
by my manual login to one of the instances, but the error message
mentioned a connection timeout on port 443. I also had problems getting
into one of the running imager instances so maybe there was a transient
network problem.

Regardless of whether it forces traffic through the private interface or
not, removing the IP as a possible attack vector is a good idea.