DEV-2188 Embed private version in image name
We have found that stripping off the timestamp portion of the version of the
public image when naming the private image causes too many problems
during development, because we can't make a second private image for the
same pipeline major/minor version without deleting the existing one.

This approach will embed the public image name and then add a second
timestamp version of the private image which will allow as many as we
need. This will also be used in tagging the private resources to make it
easy to resolve the resources used for a given image.