DEV-2358 Orange upgrade to 1.4 (#342)
DEV-2358 Orange upgrade to 1.4

Version bump plus a couple new resources.

Also implemented a fix here to check for the existence in GCS of the feature plot. This enables use
of the Orange 1.2 feature to pass a non-existent cuppa feature plot as an argument (the feature plot
is not always used).