DEV-2382 Fix some bugs in the private imager
There was an underscore in the name of the private image, which isn't
allowed under the naming rules for GCP images.

Also there was what appears to be a timing issue with the `git` checkout
from the private image VM, which was failing repeatedly from multiple
machines for us today. A simple `sleep 60` seems to have at least
allowed a single operation to succeed so let's wait and see whether that
lasts.