DEV-2492 Add new labels and labelling for disks and buckets
Introduces a couple new labels: cost_center, which can be used to bucket any cost;
user, the instigating user which will be defaulted when run locally; sample, the
tumor sample name or ref if tumor not available. Run ID also changes, removing
superfuous information from the RunTag (so run_id argument is used exactly).

Labels are refactored to be used both in the compute and storage contexts and
also in the batch framework consistently.