DEV-2547 Allow an imager instance per user
This is better than the current situation in which we can only have one
instance total. Currently when a second user attempts to start up an
imager if another is in-progress the second user will have to manually
delete the existing one.

The existing simple approach has its benefits but a couple of times
recently we've had image sessions get aborted. This new way will let one
imager run per user and does carry increased risk of machines being left
running. However these machines are not very expensive and the
improvement seems worth it.

Also improve the descriptions so it's easier to see without logging into
a machine whether it can probably be cleaned up.