DEV-2546 Add support for germline and tumor only mode
Changes to the stage interface to allow for the definition of different commands in different modes, tumor only,
germline only and somatic. The change is backwards compatible and the old Stage commands method will by
default be used at the somatic commands.

The mode is resolved based on the input, and currently only functions using JSON input.

The SomaticMetadata is modified to make the reference sample also optional, and the clients of that class are
modified as well.