DEV-2564 Publish event for verification
A second attempt at this, with fewer changes in-pipeline. The loader will bear
the brunt of the changes.

With this approach a new command-line argument has been added which will govern
the publishing of the `pipeline.completed` event when in "local" metadata mode.
If this argument is given the StagedOutputPublisher will be invoked to publish
the event; it is the same event that gets published in the regular RESEARCH and
DIAGNOSTIC contexts when an actual API is involved.

The `genomic-db-loader` is being changed to directly load the event which is
usually processed by the `archiver`. We will assume that pipelines will only be
run with this flag in an environment where it is appropriate, to avoid having
the `archiver` intercept these events where not desired.