DEV-2599 Port GRIDSS to new input modes
Biggest change is to the Driver class, which now takes the tumor and reference sample names and
paths in a builder style.

The driver is configured to pass different permutations of input bams (last argument) based on which
mode we're running.

Renamed the stage and output to Gridss and Gridss to move towards the more explicit naming found
elsewhere in the pipeline.