DEV-2596 Refactor gripss somatic/germline to consolidate shared logic
Follows the SAGE structure moving the differences between somatic and germline out into
a configuration class.

In preparation for the germline/tumor only changes we create a new builder for the gripss
command. Cobalt and Amber commands are changed to JAR commands, turns out they always had
an entry point defined.

As of this change tumor-reference is still the only mode enabled as GRIDSS is still
disabled in both modes.