DEV-2686 Write pilot images to separate family
Really a small change just to make sure that there is a pilot image
family. The original changes to add the pilot imaging capability were
done so that there is little chance of a pilot version getting beyond
the development environment.

However this change makes it easier for multiple devs to work at the
same time without trampling one another. While we've always had the
situation where multiple imagers working at the same time will trample
one another, it's never really been an issue. However the pilot workflow
and the "regular" dev workflow are distinct so it's better to keep them
separate.

This change puts the onus on the developer to make sure they're setting
the image in their run's arguments for the pilot flow.