DEV-2601 Run automated code cleanup
The main culprits that have been repaired are unoptimised imports and missing
final keywords (but not within methods).

Also there was a pair of existing test failures that have been resolved.

This would not be worth putting in a PR, it would be very difficult to review
and doing so would not provide much value.