DEV-2705 Allow private image creation from tags
This change will support the creation of images based upon a non-HEAD
commit of the private resources. It is required to allow us to do the
truthset runs for the CKB automated updates capability.

The private image creation script will continue to work as it always did
without specifying the additional arguments.