DEV-2705 Support bucket-based completion detection
This change for the private imager continues to support the existing
mode of operation, but adds support for a new flow in which instead of
using SSH to poll for activity, the script waits for the presence of the
given path in a bucket.

With this change we can support launching the imager from within a
Docker container launched by the Cloud Build environment and not having
to fiddle with creating and propagating SSH keys to the VM instance
doing the imaging.

When the script is provided a bucket URL argument, it switches into the
new bucket-based mode. In that case instead of copying the private
resource checkout script to the imager instance after it starts and
waiting for it to complete, the script will set the startup script
metadata of the instance to the local file. Also instead of passing
arguments on the command line we set metadata on the instance so it can
be retrieved without an SSH connection.

In addition, argument parsing has been modified in both the
`create_private_image.sh` and `private_resource_checkout.sh` scripts to
move away from positional parameters and to `getopt` long options for
clarity and also to support the new capabilities, as purely positional
parsing wouldn't support all cases with the new optional arguments.