DEV-2705 Add starting point for Protect-only runs
This can probably use some work as several of the later steps are run, not just
Protect. Also a couple of the recently-added stages were never added to the
existing starting points so I have added those.

I have also added a `Namespace` annotation for the stages and a couple of tests
that use reflection to do some basic checks on the namespaces and starting
points. This may be a first step toward a more cohesive structure for the
starting points but also is a move closer to a more official interface for the
stages. For the moment almost all of the stages reference existing constants
as the value for their new annotation but this too is just a first step.