DEV-2705 Support persisted output for Lilac
This was required to allow us to use the persisted output for Lilac rather
than running it as part of the Protect runs (itself intractable as Lilac
cannot rerun from CRAM output currently).

Also fixed a little bug in the Purple persisted output which was causing
failures in the Protect and Orange steps (and maybe others I had already
removed from the Protect-only starting point). The trouble here was that
the being used for a trio of germline files was hardcoded to the reference
sample rather than the "metadata" sample, which does the right thing depending
upon the mode (tumor-reference, reference-only, tumor-only).

Finally removed the remaining stages from the Protect-only starting point.