DEV-2767 Read lane names from the API
Previously these would be parsed out of the filename.

It feels to me like we should consider using the Java API client for this
and the other interactions in the pipeline but I'm not sure if there is a
reason for not moving away from what we've got. For now I've extended the
roll-your-own approach but would prefer to switch to the client soon if
there is no compelling reason not to.