DEV-2778 Bump Purple memory
On our recent 5.28 reruns there were a few samples that failed due to the Java
heap running out.

Just double the memory on the box itself and increase the heap so that it
leaves the same amount of non-Java memory free on the box (4GB).

Also the version of the pipeline is currently being read out of the run
rather than the pipeline artifact to allow us to use triplet versions for
the moment. This should probably not be propagated if the branch is merged.