DEV-2793 Allow using branch for public resources
This makes our development process more robust by allowing multiple developers
to work in parallel on tickets that amend the resources, and also protects the
production image from changes that haven't been explicitly merged to `master`.

When a branch is used for the public image, a distinct family will also be used
and no tag will be written to the repository so that image won't be available
to use as the base for a private image.

I've also amended the private imaging script to allow passing any "target"
instead of just a SHA on the command line; the user can now use a tag, commit
SHA or branch. When run from within GCP though only a SHA will be supported.

Options in the public script have switched to long-only to match the private
script's behaviour.