DEV-2798 Python venv support for pilot images
We create Python virtualenvs in their own directories for several of our
tools. However when we try to run these tools with a pilot version,
there is no virtualenv with a name matching the version and so it can't
be found at runtime.

I've added a script to locate the virtualenv for the latest version of
the tool (with some limitations) and this should be enough to get us
working in 99% of cases. Actually I expect this will only ever really be
an issue with Cuppa but it's possible we add tools later in a similar
vein.