DEV-2705 Allow private image named for SHA1
This makes the whole process of locating and managing the images for the
Protect differencer simpler. With this change the build that
automatically creates the Protect imager can pass the SHA1 through to
the other build that does the actual differencing without jumping
through hoops trying to determine the name.

For now it's explicitly for this process so the arguments require the
commit to be given if using this option but I can see it generalising
nicely to the rest of the pipeline.