DEV-2831 Add in new arguments for primary tumor
These are currently required in ROSE, but the curated primary tumor file is
clinical data that comes in as a TSV. This is a temporary change to avoid
having to make another ROSE release.

We anticipate that ROSE will become part of the forthcoming reporting pipeline
which will have some different mechanism for dealing with clinical data used
as input.