DEV-2853 Check for correct getopt before running
This may be an anaemic check but it is better than what we had. On MacOS
the default getopt implementation is based on the BSD version and
doesn't support long options.

When invoked with "-V", the default version on modern Linux
installations (from the GNU foundation) returns a version, while the
MacOS version just returns "--". I've used this to create a simple
suitability check and we now print a warning with details for MacOS
users.