DEV-2974 Add "publish events only" mode/argument
This will allow us to escape from the current situation in which the events for
the 5.30 rerun have expired leaving us with no way to archive them now that
they have been verified.

By invoking the pipeline with the new argument set we should incur little cost
and also be more-confident of the correct outcome than if we were for instance
trying to construct the datasets out of the data in the output bucket.

Also fix the `SageCommandBuilderTest` which appears to have been broken since
the last commit to the hotfix branch.