DEV-2991 Handle GCP InvalidArgumentException
We have witnessed this exception multiple times in production but it does not
reproduce easily. Because we retry forever when there is any issue this bug
threatens the stability and turnaround characteristics of the pipeline, since
once it has been encountered once in a given invocation it will never succeed.
We have also seen it in the batch framework.

The current handling mechanism just prints out the message from any exception;
I have decided to abort when this specific exception is encountered and instead
of retrying print the full stack trace. If nothing else maybe we will get hints
on what is happening so we can address the root cause. This will also ensure
the pipeline doesn't sit forever but instead the rest of our framework will
restart the job.

In my testing this approach results in a proper error message being printed
and from what I can tell a proper death of the thread containing the retry
loop. Hopefully this holds in the wild but I'm limited in how well I can unit
test this code while still retaining a realistic representation of how the
GCP libraries behave in use.