DEV-3052 5.31 rerun starting point
This is not pretty code but these rerun starting points do not live
forever. If there is some slack time during the rerun I may come back
and revisit to see if we can have a nicer interface.

Also improve the help for the starting points which was outdated and
thus incorrect.