Increment Compar version to shaded JAR
This was un-done not long ago to deal with classpath issues on some
client machines but appears to becausing problems with build.

I am not sure whether the problems will be resolved with this because
this new version is built from a different point in the repository than
the previous versions. All those in Maven have been unofficial so far
because a "real" tag was only recently made and we were early adopters
in the pipeline.

A new version is necessary because we are now using the Artifact
Registry rather than the bucket-based `hmf-maven-repository` repository,
into which one could overwrite files if desired.