DEV-3089 GRIDSS change from master to 5.31
David identified a problem in the 5.31 validation against the 290 samples
we had chosen to test svprep and Charles has recommended pulling his changes
for the forthcoming release into 5.31.

The fix reorders the BAMs and arguments to the driver as this ordering affects
Purple. Going forward Purple will not be sensitive to ordering but for now this
is the pragmatic thing to do.