DEV-3113 Bump sv-prep to 1.1
We have decided to do this as we encountered 3x samples in the production
research environment which got “stuck” on sv-prep  and never finished. These
samples have (in one case only moderately) more reads than is typical. This
new version of sv-prep significantly improves memory efficiency and we have
decided that this is something we want to have in place before we update the
diagnostic environment to 5.31.

This change involves a version bump in the pipeline and a new disk image
containing the new version.