Force JAR rebuild when imaging
The more I thought about this the more I thought it made sense not to
assume the JAR is up-to-date.

I did it a bit more cleanly than the first time I tried which involved
changing to the directory and issuing the Maven command, this doesn't
involve any context change whatsoever.