DEV-3166 Revert changes to datatype constants
Several datatypes were renamed as part of DEV-3080 but weren't strictly part
of that change, they were just done in the same commits.

There is quite a lot of hassle involved with renaming datatypes, and it is not
obvious to me that these changes would have been worth that hassle. We should
come back to this as part of a considered refactor in the future.