DEV-3178 Further simplify VM management
I have removed the messy loop that previously composed much of the
executeSynchronously() method, replacing it with the mechanism recommended in
Google's documentation for waiting on a running operation.

Logging of any errors in the InstanceLifecycleManager now occurs at the ERROR
level, though most exceptions will be thrown back up to the StageRunner where
the stage itself will be retried. This is somewhat of a middle ground; it can
only result in more production alerts but should help us get a handle on which
operations actually routinely fail, while also properly handling any errors
that do come up.

There is more work to do here but I am reluctant to do too much as this is such
a core piece of our infrastructure. I will revisit at some point in the future,
when this release has been out in the wild for a while.