DEV-3178 PR suggestions and bug fix
I noticed that there is a problem actually removing old succeeded VMs, that
path is a bit tricky to hit in testing just due to timing. I believe that is
now fixed as well as the requested changes from the PR.

Also made the ordering of arguments more consistent within the instance
lifecycle manager class, it was getting slightly confusing.