DEV-3178 Clean up zone handling for deletes
The zone returned from the instance is a full URL, but the deletion expects a
short zone like "europe-west4-a". Previously this was being sanitised in the
`deleteOldInstancesAndStart` method and I just added it to the `delete` method,
but they were doing it in different ways. I have moved it into the `delete`
method which is now called by `deleteOldInstancesAndStart`, thereby removing
the duplication.