DEV-3234 Fix bug in old instance deletion
The deletion code was incorrect because it located an old instance but then
tried to delete that from the zone the new instance was going to be placed
into. This would work sometimes (as there are only 3 zones in europe-west4)
but not always.

With the recent change to the logging it is now obvious what was happening.