DEV-3240 Pass folder at composition and some refactoring
Refactoring = rename everything "report" to "output" as the report is an old legacy name before we
understood really what the patient report was.

Also rename the PipelineResults to PipelineOutputComposer which is more true to its purpose.

Move the Noop stuff out of a boolean and into a Noop implementation of the composer.

Remove as much unnecessary code as I could, including the qualifier which was actually a
temporary debug thing from some time ago. We use the folder instead to draw the distinction
between somatic tumor and ref.