DEV-3242 Fix public image creation
We discovered yesterday that we could no longer properly install the packages
for Debian9 as they have moved them to an archive server and it seems have
also changed some of the versions. I tried to get everything working using
version pins and quickly found myself in dependency hell, and it just wasn't
going to work.

Instead I created a VM from the last-successful image and deleted all the
customisation (essentially just the files from ``/opt`) and cut a new image
from that in the GCP console. I applied the family `hmf-debian-9` to that
and updated the `create_public_image.sh` script to reference it instead.

I've also commented out everything in the `*.cmds` files, I think this is fine
(ie as opposed to deleting them) as we are likely to be dredging all the
commands out again in a month or two anyway.