DEV-3243 Add temporary research starting point
We recently disabled the germline caller in research, moving it to diagnostic.
The research launcher copies the germline output along with the alignment
output from the diagnostic output bucket as it always did. However since the
germline stage is skipped in the research run, we don't mark the output for
that stage with any datatypes. Using this new starting point will result in
the datatypes for the germline being applied which should correct the problem.

A corresponding change will need to be made in the launcher code and this will
fix the problem for forthcoming runs. There are some runs that have already
been archived that will not be fixed, I will craft an SQL file to update the
API file entries for these as the only thing missing is the datatypes (ie the
archiver otherwise ran properly).

This research starting point can be removed after the next diagnostic upgrade
when we will stop running the germline caller in the diagnostic pipeline.