DEV-3329 HMF Tools updates for the 5.33 release
Upgrades the full suite of hmftools for the next major diagnostic pipeline release. See the release notes in 
sharepoint for full details of the changes and validations.

The Versions file is now split into two classes with more information about the tool configuration: HmfTools and
ExternalTools. 

Rose and Protect are removed from the pipeline. 

Compar is upgraded to 1.1 and provides more comprehensive regression in the smoke test.