DEV-3508 Refactor starting points for 5.33 rerun
The desired steps for the 5.33 rerun are the same as the set for the 5.32 were
so I decided to just make a new permanent starting point for this situation,
and have called it `PAVE_GRIPSS_ONWARD`.

I noticed there was quite some duplication between some of the existing sets
so eliminated that. I also decided to just leave in the non-germline Orange
invocation that was added for ACTIN as Orange runs very quickly and it seemed
likely the ACTIN pipeline might be rerun at some point in the future.