DEV-3688 Unify VM core and memory declarations
The `HmfTool` class and the VM definitions both define parameters for the VMs
that our in-house tooling runs on. I noticed this while trying to increase the
memory for Purple and when it wasn't working realised the situation.

Reference the HmfTool-hosted constants from within the VM definition calls
whenever they exist as not doing this violates the principle of least surprise.
Also I think the intent of the HmfTool class must have been to centralise the
configuration as much as possible and this is a step in that direction.

I am aware this is an imperfect solution as the constants don't exist for those
tools we don't maintain and a couple of the tools use different parameters
depending on which "mode" they've been called in, Sage springs to mind.

I checked pretty thoroughly for references and I think I got them all but I
suppose we'll find out in the SmokeTest if anything is off.