DEV-3720 Imager changes to use Artifact Registry
In the fullness of time this will replace the `common-tools` bucket
entirely for HMF-maintained tools. For the moment it simply provides an
alternative flow to allow fetching artifacts from our internal Artifact
Registry. As tools are made available inside the registry they can be
removed from the bucket and added to the list in `VersionUtils` to be
fetched from the registry instead.