DEV-3720 Fix permissions errors on artifact fetch
The changes here are small but they are the result of quite some work.

First, the default scope was not enough to fetch objects from the
Artifact Registry so I've changed the VM to have all scopes. The IAM
permissions and the scopes work together to provide the configured
access so this could not be done with IAM alone.

Second, the version of `curl` on the build VM (which is based on an
older Debian 9 release) does not insert the same header values when
invoked with the original `--oauth-bearer` as the version that must have
been used to test the command before I got it, and indeed for all
testing I did of it (all on newer versions). So just explicitly set the
appropriate header based on what the newer versions set, which works.