DEV-3720 Move imager to top level
The imaging scripts I think are better positioned at the top-level of
the project since they are so important to having a functioning product.

Note that the `base.cmds` file in the past had much more content which
was removed when Google deprecated the source `debian9` image and
replaced with a comment. I think that original content will now be
harder to find but would likely come in handy when migrating once again
to a supported upstream image.