DEV-3720 Move registry fetch to its own script
Putting the commands in a script allows more control over failure
scenarios and eliminates the painful quotation histrionics since the
commands were generated locally and pushed over the wire previously. Now
all that is required is the passing of the tool name and version to the
script on the remote instance. As with a couple of others this script is
just copied over to the imaging VM at startup.

Also correct a couple of bugs to do with the previous commit that moved
the scripts to a top-level directory in the source tree.