DEV-3730 Dredge up old imager commands
These had previously been removed because the official Debian 9 mirrors
had been discontinued upstream and the packages could no longer be
installed. When this happened I had "froze off" a new parent image from
the previous successful imaging run and we used that for our parent for
the pipeline5 imaging instead.

Now these are again needed as this ticket is about fixing the imager to
work with Debian 12 now.

Also make some improvements to the imager itself (needed as a static
timeout was no longer good enough to avoid the VM not being reachable)
and update the `mk_python_env` script somewhat as Debian 12 now includes
Python 3.x instead of Python 2.7.