DEV-3730 Remove Python version from mk_python_venv
Previously commented out inside the script, I removed that commented-out
line and also removed it from the CLI arguments.

I anticipate with a more-modern Python as the base install that we
shouldn't need to be able to provide the Python version on the command
line, this was essentially always just a 3.x version before when the OS
bundled Python 2.7.