DEV-3730 Amend requirements.txt version ranges
This is perhaps a bit of an overstep for the imager but will allow it to
work at least for a while. Principally due to the fact that pycuppa is
under active development and we don't know where it will land, this
change should make it possible to install the application on the newer
Debian version without insisting on exact versions.

Also the Python script now installs the entire `requirements.txt` as a
unit which is different from one-at-a-time as it used to do. The
previous behaviour was required way back but clearly not anymore.