DEV-3730 Fix broken GATK setup
The GATK versions we're using require a 1.8 JDK, I believe this is to do
with classloader implementations and how that is used to find the actual
algorithm that is being requested.

Perhaps these changes are ugly enough that we would prefer to look at
updating the GATK versions instead or exploring some other solution but
this approach does appear to work.