INFRA-190 Bump Cuppa version in the pipeline
Also remove Cuppa from the assertions that will be done by the smoke
test Compar invocation as Charles indicated via email this will not work
properly with the new Cuppa version. I expect this will significantly
change the expected output so there should be more changes there but
that won't be until we get the whole thing hanging together.