INFRA-190 Update to Cuppa 2.0
These changes were already made by Luan in AUS-204 but the changes in AUS-205
were done in parallel and essentially both are required for 5.34. Given the
AUS-204 changes are just for Cuppa I decided to try to integrate them overtop
the AUS-205 branch.

In addition to Luan's changes there are modifications to the imager here that
support the module-style execution for Cuppa 2.0 which is different from
previous Cuppa versions (or rather the `cuppa-chart` Python code shipped
inside previous Cuppa JARs). Also new R libraries were needed and a new
Python version to match the one used for Pycuppa testing. Also a couple of
DEB packages have been pulled in.

I've tried to retain everything I could from Luan's original commits but one
thing I had to change which was the use of a `SubCommand` to wrap the compound
`Python3ModuleCommand` so that stderr/stdout are both retained in the logs.
Also I replaced `;` delimiters in the `Python3ModuleCommand` generated code
with `&&` since this is actually more correct and supports the expected failure
mode in case there is any problem with the virtual environment.

At this point Cuppa is properly called and does run but has some errors in its
output, I may have simply missed something when porting the AUS-204 changes.