INFRA-190 Go back to "proper" image name
This was changed to allow me to test without causing any problems for anybody
else; now that the Cuppa 2.0 changes are incorporated properly I'm happy for
it to go back.

Also update the Orange version to a beta JAR (built from hmftools commit
93e5b9cf5cf6c8bee032260b3a256e87251b8ee0) in lieu of Luan's official version
which is likely to come early next week and fix up formatting on the Cuppa
class.