INFRA-190 Cleanup per PR suggestions
I elected to remove the not-strictly-related but definitely unreferenced
Bachelor datatypes along with the old Cuppa ones.

Overwriting the R libraries tarball would have broken older imager scripts
so I made a copy with an embedded date since we are likely to be doing this
again soon (the Debian 12 upgrade is waiting in the wings for 5.34 to go out).