INFRA-190 Use diskimager net/subnet for SmokeTest
I am sure this will be something we want to change pretty quickly but for now
it should work as there is plenty of room on the subnet for a single pipeline
run.

This was necessitated by the removal of the default network from the
hmf-pipeline-development project. I have elected to make the change here rather
than in the `master` branch so we can get to working builds for this and
the AUS-205 branch sooner.