DEV-3730 Imager changes for Debian 12
The `base.cmds` were pulled out of the history in source control to use
as a starting point for going from the Google-supplied image to our base
environment. I've deleted the second `cran40` R installation script
because it should no longer be needed once this upgrade is complete.

Also `mk_python_env` has been updated to work with the Python3
environment that is now the default in Debian 12.