DEV-3730 Pull in Anaconda-created R library
This is a bit of a hack just to keep momentum going here, essentially for each
command that needs an R environment I am calling the Anaconda environment
enablement commands inline. If I can verify that the various steps of the
pipeline all work with this approach then I will find a nicer way to do it.

Also include the AdoptiumJava8JarCommand above those tools using
GATK because I had to go to lengths to get a Java8 version installed in Debian
12 and this is all a bit speculative still. Again if I can get everything
working I will find a nicer way to do this, though I haven't quite decided
if this class really belongs in `hmf-cloud-sdk`.