DEV-3730 Artificially bump GRIDSS-related tools
These are not new versions of GRIDSS or Virus Breakend but rather recompiled
binaries for the existing version. Yet again this is a temporary approach to
see if we can get a completing pipeline.

Important to note that the latest version for GRIDSS on Github is actually
2.13.2 but there are later commits and we are already using 2.13.3 in the "old"
pipeline; I assumed this was just built from the later commit but that is
untested. It is not wise in my opinion to bump the tool version when we have
just rebuilt the same code for a different environment so we need to come
up with a strategy here, but I am reluctant to do much engineering as this
whole approach (ie GCP disk images) may not have long to live, so I'm just
documenting and drawing attention to what has been done.