DEV-3730 Disable disk allocation statements
This is actually related to the changes I've been making in parallel in
the `hmf-cloud-sdk` project, which replace the "|| die" pattern at the end
of the generated BASH commands, with trap-handling functions.

As part of the changes I decided not to exclude the machine startup commands
that create the software RAID arrays from the handling and immediately saw
failures. I don't know exactly the root cause yet but suspect that the custom
machine types have changed since this code was written and now adding a single
local SSD is disallowed for the given machine parameters. I do believe the
failures were to do with the local SSDs based on forensics I did on the failing
instance. Not sure when this stopped working. Or maybe it never worked, we have
no way of knowing.

Rather than remove the code entirely I've left it because I am just pushing
to get the pipeline working. The code should work because I think now we will
get 4x local SSDs rather than 1x, so no new failures will happen. However this
is not optimal, I should have time to come back and figure out what these
stages should specify if not 1x local SSDs.