INFRA-250 Add pipeline5-specific startup commands
These are now supported by `hmf-cloud-sdk` and will appear in the preamble of
the generated BASH script that is executed on the VMs. There is much less
duplication using this approach than there would be doing it in the object
hierarchy since much of that is hosted in the `hmf-cloud-sdk` project and
having to compose it into every stage that needed it didn't seem like a
worthwhile tradeoff for the improved documentation that would offer.

I strongly believe this is better than pushing pipeline5-specific code into
the common library.