INFRA-250 Cleanup imager and README
Try to capture what was done to get Debian 12 working with Anaconda in
the README and leave only what is still in use.

Also add the `anaconda.yaml` file which is the result of a `conda env
export` from the builder machine. This should help whomever comes in to
do this next.