INFRA-250 Remove old Chord R script
This had not been used for some time so far as I can tell. Also I
removed the calls to copy the "profile" files from my tests with the
virtual environment setup, which have been replaced with the changes
that allow "startup" commands to be specified for all pipeline stages.

Also remove the "ned-" prefix from the name of the pipeline image and
just clean up some of the documentation in the scripts.