INFRA-250 Replace minimum disk-space parameters
I've returned these now with "0" as the smallest supported size as we have been
running this way for some time (unbenownst to us) already.

Also add in the `unzip` dependency to the imager that was being pulled in
transitively before my last cleanup and remove the verbosity from R environment
setup.

Remove the final few commands for R initialisation in a couple of the stages
that I had forgotten about.